pyproject.toml
[project]
name = "ansmetheus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paramiko>=4.0.0",
    "pyyaml>=6.0.3",
]
